[v0.91.2][maintainability] Split uts_acc_multi_model_benchmark by responsibility
